+++ src/main/java/kr/hhplus/be/server/global/config/SwaggerConfig.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.common.config.swagger;
+package kr.hhplus.be.server.global.config;
 
 import io.swagger.v3.oas.models.OpenAPI;
 import io.swagger.v3.oas.models.info.Info;
